Move LogEvent to api package
