+++ jib-core/src/main/java/com/google/cloud/tools/jib/api/LogEvent.java
@@ -14,9 +14,8 @@
  * the License.
  */
 
-package com.google.cloud.tools.jib.event.events;
+package com.google.cloud.tools.jib.api;
 
-import com.google.cloud.tools.jib.api.JibEvent;
 import com.google.common.annotations.VisibleForTesting;
 import java.util.Objects;
