+++ jib-core/src/test/java/com/google/cloud/tools/jib/api/MainClassFinderTest.java
@@ -17,7 +17,6 @@
 package com.google.cloud.tools.jib.api;
 
 import com.google.cloud.tools.jib.api.MainClassFinder.Result.Type;
-import com.google.cloud.tools.jib.event.events.LogEvent;
 import com.google.cloud.tools.jib.filesystem.DirectoryWalker;
 import com.google.common.io.Resources;
 import java.io.IOException;
