+++ jib-plugins-common/src/main/java/com/google/cloud/tools/jib/plugins/common/logging/ConsoleLogger.java
@@ -16,8 +16,8 @@
 
 package com.google.cloud.tools.jib.plugins.common.logging;
 
-import com.google.cloud.tools.jib.event.events.LogEvent;
-import com.google.cloud.tools.jib.event.events.LogEvent.Level;
+import com.google.cloud.tools.jib.api.LogEvent;
+import com.google.cloud.tools.jib.api.LogEvent.Level;
 import java.util.List;
 
 /** Logs messages to the console. Implementations must be thread-safe. */
