+++ jib-plugins-common/src/test/java/com/google/cloud/tools/jib/plugins/common/logging/ConsoleLoggerBuilderTest.java
@@ -16,7 +16,7 @@
 
 package com.google.cloud.tools.jib.plugins.common.logging;
 
-import com.google.cloud.tools.jib.event.events.LogEvent.Level;
+import com.google.cloud.tools.jib.api.LogEvent.Level;
 import com.google.cloud.tools.jib.plugins.common.logging.ConsoleLoggerBuilder.ConsoleLoggerFactory;
 import com.google.common.collect.ImmutableMap;
 import java.util.ArrayList;
